requirements.txt
djangorestframework
python-decouple
celery
requests
psycopg2-binary
dj-database-url
redis